#28 add owner update functionality (#50)
Owner update functionality

an extra panel has been added for updating owner information
updating (and saving the new information) is stored in the authentication service. When we decouple user from owner, this has to be redone, as it is rather sloppy now. Then, it will be refactored into separate authentication, user and owner services