Change default template lang to en (#115)
* change default lang for angular templates to en

* make changes for i18n and l10n

## env.js
Standard Angular i18n tools are used for delivering multiple languages. These
language are available under, i.e. /en/, /nl/. However, this breaks the way we
define the API endpoint, i.e. /api. As we've got multiple clusters, we don't
want to build a separate image for each backend.

For this reason, an env.js has been introduced. The variables we define there
can be overridden in the entrypoint.sh, which is run once the image is deployed.
Other variables can also be stored here and set as environment variables in the
future instead of the Angular environment file. That one is still needed for
enabling production mode.

## switching language
A language switcher component has been added to provide links to each language.
There is no sophisticated mechanism to determine browser language, or store
language preferences. We might want to expand on this in the future.

## nginx
Dutch is now shown as default. Additionally, the user is redirected if they try
to visit /nl/ pages, i.e. domain.com/nl/owner -> domain.com/owner. This makes
sure that Dutch is as much default as possible.

* incorporate sonarcloud feedback

* fix release script

* fix import