Prevent multiple identical devices (#197)
* refactor: split saveDevice in two functions

* fix: submit form only one at a time

when pressing Save multiple times at once, the form was sent multiple times to the back-end. Lock the request to only submit once by setting this.submitting

* feat: set device form to edit mode once new device has been created

This way, the device is only created once. Subsequent Saves just update the device

* feat: wait w/ editting for device to become available

not all CQRS views might yet be updated after the device was succesfully registered. This merely means that the Events were created. For editing the device, the device needs to be available in the Mongo projection. If it not exists yet, wait for it

* refactor: fix some types

* refactor: add subscription to watch list, so we can unsubscribe